对后面进行数据处理的部分，分成了多个不同的函数线程用来订阅下位机单个子系统中的不同长度的数据

+++ dataservice/zmq/UPcomputer_part/data_process_0mqsubsys/data_process_11_acceleratepower_num2.py
@@ -5,7 +5,7 @@ slave：11
 port:5001
 
 子系统需要检测的信息
-电源电压采样 value1:05 03 07 data crc1 crc2----registerid=07   datatype=float
+电源电压采样 value1:05 03 07 data crc1 crc2----registerid=07   datatype=int
 电源电流采样 value1:05 03 08 data crc1 crc2----registerid=08   datatype=float
 
 '''
